#5 update: 팔로워 수, 팔로잉 수 데이터 타입 변경

+++ src/main/java/sg/graphServer/entity/Account.java
@@ -20,7 +20,7 @@ public class Account {
     private String accountName;
     private String profileImage;
 
-    private int followingCount;
-    private int followerCount;
+    private long followingCount;
+    private long followerCount;
 
 }
